+++ example/lib/app.dart
@@ -1,6 +1,7 @@
-import 'languages/languages.dart';
+import 'languages/language_dart.g.dart';
+import 'widgets/widget_dart.g.dart';
 import 'package:finch/model.dart';
-import '../core/local_events.dart';
+import 'core/local_events.dart';
 import 'package:finch/finch_route.dart';
 import 'db/job_collection_free.dart';
 import 'db/person_collection_free.dart';
@@ -15,20 +16,21 @@ import 'route/web_route.dart';
 import 'package:finch/finch_capp.dart';
 
 FinchConfigs configs = FinchConfigs(
+  jinjaMapTemplate: mapTemplates,
   widgetsPath: pathTo(env['WIDGETS_PATH'] ?? "./lib/widgets"),
   widgetsType: env['WIDGETS_TYPE'] ?? 'j2.html',
   languagePath: pathTo(env['LANGUAGE_PATH'] ?? "./lib/languages"),
   languageSource: LanguageSource.dart,
-  dartLanguages: languages,
+  dartLanguages: languageDart,
   publicDir: pathTo(env['PUBLIC_DIR'] ?? './public'),
   dbConfig: FinchDBConfig(
     enable: true, //env['ENABLE_DATABASE'] == 'true',
-    dbName: 'example',
-    auth: 'admin',
-    pass: 'PasswordMongoDB',
-    host: env['MONGO_CONNECTION'] ?? 'localhost',
-    port: env['MONGO_PORT'] ?? '27018',
-    user: 'root',
+    dbName: env['MONGODB_NAME'] ?? 'example',
+    auth: env['MONGODB_AUTH'] ?? 'admin',
+    pass: env['MONGODB_PASSWORD'] ?? 'PasswordMongoDB',
+    host: env['MONGODB_CONNECTION'] ?? 'localhost',
+    port: env['MONGODB_PORT'] ?? '27018',
+    user: env['MONGODB_USER'] ?? 'root',
   ),
   port: (env['DOMAIN_PORT'] ?? '8085').toInt(def: 8085),
   mysqlConfig: FinchMysqlConfig(