+++ lib/src/finch_app.dart
@@ -1,6 +1,9 @@
 import 'dart:async';
 import 'dart:io';
 import 'package:capp/capp.dart';
+import 'package:finch/src/cli/commands/commands.dart';
+import 'package:finch/src/tools/convertor/widget_to_dart.dart';
+import 'package:finch/src/tools/convertor/language_to_dart.dart';
 import 'package:mysql_client/mysql_client.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:finch/src/db/mysql/mysql_migration.dart';
@@ -168,7 +171,7 @@ class FinchApp {
 
     await mongoDb.close();
     await mysqlDb.close();
-    sqliteDb.dispose();
+    sqliteDb.close();
 
     _mongoDb = null;
     _mysqlDb = null;
@@ -241,13 +244,27 @@ class FinchApp {
 
     server!.address;
 
-    Console.p({
-      'url': config.uri,
-      'path': pathApp,
-      'DB': {
-        'path db': config.dbPath,
-      }
-    });
+    CappConsole.writeTable(
+      [
+        ['', '@> Finch ${FinchApp.info.version}'],
+        ['Url', config.uri.toString()],
+        ['Path App', pathApp],
+        if (config.dbConfig.enable && _mongoDb != null && _mongoDb!.isConnected)
+          ['MongoDB', 'Connected'],
+        if (_mysqlDb != null && mysqlDriver.connected()) ['MySQL', 'Connected'],
+        if (config.sqliteConfig.enable &&
+            _sqliteDb != null &&
+            sqliteDriver.connected())
+          ['SQLite', 'Connected'],
+        if (hasSocket) ['WebSocket', 'Enabled'],
+        if (config.isLocalDebug) ['Debug Mode', 'Enabled'],
+        if (crons.isNotEmpty) ['Cron Jobs', crons.length.toString()],
+        if (commands.isNotEmpty) ['Has Commands', 'Yes'],
+        if (appLanguages.keys.length > 1)
+          ['Languages', appLanguages.keys.join(',')],
+      ],
+      color: CappColors.info,
+    );
 
     await handleRequests(server!);
     if (awaitCommands) {
@@ -280,6 +297,7 @@ class FinchApp {
                 });
               } else if (!_mongoDb!.isConnected) {
                 await _mongoDb!.open().onError((error, stackTrace) async {
+                  Console.e("Error connect to DB: ${config.dbConfig.link}");
                   throw ("Error connect to DB");
                 });
               }
@@ -330,6 +348,16 @@ class FinchApp {
     await _runCommands(_args);
   }
 
+  final _helpOption = CappOption(
+    name: 'help',
+    shortName: 'h',
+    hideInHelp: true,
+    onSelect: (CappController controller) {
+      controller.writeHelp();
+      return false;
+    },
+  );
+
   /// Creates and configures the command manager for interactive CLI operations.
   /// Sets up a [CappManager] with all available server commands including:
   /// - `help` - Shows help information for all available commands
@@ -352,10 +380,15 @@ class FinchApp {
         ),
         args: args,
         controllers: [
-          CappController('clear', options: [], run: (c) async {
-            CappConsole.clear();
-            return CappConsole.empty;
-          }),
+          CappController(
+            'clear',
+            options: [_helpOption],
+            description: 'Clear console',
+            run: (c) async {
+              CappConsole.clear();
+              return CappConsole.empty;
+            },
+          ),
           CappController(
             'help',
             options: [
@@ -366,15 +399,13 @@ class FinchApp {
               ),
             ],
             run: (c) async {
-              return CappConsole(
-                c.manager.getHelp(c.manager.controllers),
-                CappColors.warning,
-              );
+              return c.manager.writeHelpModern(c.manager.controllers);
             },
           ),
           CappController(
             'migrate',
             options: [
+              _helpOption,
               CappOption(
                 name: 'init',
                 shortName: 'i',
@@ -407,7 +438,7 @@ class FinchApp {
                 description: 'List migration',
               ),
             ],
-            description: 'Migration commands',
+            description: 'Mysql/Sqlite Migration commands',
             run: (c) async {
               if (c.existsOption('init')) {
                 var res = await CappConsole.progress<List<String>>(
@@ -455,8 +486,7 @@ class FinchApp {
 
                 var res = await CappConsole.progress<String>(
                   "Creating migration...",
-                  () async =>
-                      MysqlMigration(mysqlDriver).migrateCreate(name: name),
+                  () async => MysqlMigration.migrateCreate(name: name),
                 );
                 return CappConsole(res);
               }
@@ -486,8 +516,73 @@ class FinchApp {
             },
           ),
           CappController(
-            'migrate_sqlite',
+            'build_widgets',
+            description:
+                'Generate widgets Dart file from templates (to dart map variable)',
             options: [
+              _helpOption,
+              CappOption(
+                name: 'path',
+                shortName: 'p',
+                description: 'Path of templates',
+                value: config.widgetsPath,
+              ),
+              CappOption(
+                name: 'extension',
+                shortName: 'e',
+                description:
+                    'File extension of templates default (*.${config.widgetsType})',
+                value: config.widgetsType,
+              ),
+            ],
+            run: (c) async {
+              var res = await WidgetToDart(
+                c.getOption('path'),
+                fileExtention: c.getOption('extension'),
+              ).generate();
+              return CappConsole("Generated file at: $res");
+            },
+          ),
+          CappController(
+            'build',
+            description: 'Build Project',
+            options: [_helpOption],
+            run: (c) async {
+              await ProjectCommands().build(
+                c,
+                copyLang: config.languageSource != LanguageSource.dart,
+                copyWidgets: config.jinjaMapTemplate == null,
+              );
+              return CappConsole("Build completed");
+            },
+          ),
+          CappController(
+            'build_language',
+            description:
+                'Generate language Dart file from language files (to dart map variable)',
+            options: [
+              _helpOption,
+              CappOption(
+                name: 'path',
+                shortName: 'p',
+                description: 'Path of templates',
+                value: config.languagePath,
+              ),
+            ],
+            run: (c) async {
+              var res = await LanguageToDart(
+                c.getOption('path'),
+              ).generate();
+              config.dartLanguages = res.map;
+              appLanguages = res.map;
+              return CappConsole("Generated file at: ${res.path}");
+            },
+          ),
+          CappController(
+            'migrate_sqlite',
+            description: 'Sqlite Migration commands',
+            options: [
+              _helpOption,
               CappOption(
                 name: 'init',
                 shortName: 'i',
@@ -515,7 +610,6 @@ class FinchApp {
                 description: 'List migration',
               ),
             ],
-            description: 'Migration commands',
             run: (c) async {
               if (c.existsOption('init')) {
                 var res = await CappConsole.progress<List<String>>(
@@ -579,7 +673,9 @@ class FinchApp {
           ),
           CappController(
             'language',
+            description: 'Language management commands',
             options: [
+              _helpOption,
               CappOption(
                 name: 'list',
                 shortName: 'l',
@@ -630,42 +726,52 @@ class FinchApp {
               return CappConsole("Language commands");
             },
           ),
-          CappController('info', options: [], run: (c) async {
-            CappConsole.writeTable(
-              [
-                ['Info', 'Details'],
-                ['Finch version', info.version],
-                ['Dart Versions', Platform.version],
+          CappController(
+            'info',
+            description: 'Display server information',
+            options: [_helpOption],
+            run: (c) async {
+              CappConsole.writeTable(
                 [
-                  'Memory Usage',
-                  ConvertSize.toLogicSizeString(ProcessInfo.currentRss)
+                  ['Info', 'Details'],
+                  ['Finch version', info.version],
+                  ['Dart Versions', Platform.version],
+                  [
+                    'Memory Usage',
+                    ConvertSize.toLogicSizeString(ProcessInfo.currentRss)
+                  ],
+                  [
+                    'Max Memory Usage',
+                    ConvertSize.toLogicSizeString(ProcessInfo.maxRss)
+                  ],
+                  [
+                    'Socket Connections',
+                    (socketManager?.countClients ?? 0).toString(),
+                  ],
+                  ['Socket Users', (socketManager?.countUsers ?? 0).toString()],
                 ],
-                [
-                  'Max Memory Usage',
-                  ConvertSize.toLogicSizeString(ProcessInfo.maxRss)
-                ],
-                [
-                  'Socket Connections',
-                  (socketManager?.countClients ?? 0).toString(),
-                ],
-                ['Socket Users', (socketManager?.countUsers ?? 0).toString()],
-              ],
-              color: CappColors.info,
-            );
-            return CappConsole('\n');
-          }),
-          CappController('exit', options: [], run: (c) async {
-            await CappConsole.progress(
-              "Bye bye!",
-              () async {
-                await stop(force: true);
-                await Future.delayed(Duration(seconds: 1), () {
-                  exit(0);
-                });
-              },
-              type: CappProgressType.circle,
-            );
-          }),
+                color: CappColors.info,
+              );
+              return CappConsole('\n');
+            },
+          ),
+          CappController(
+            'exit',
+            description: 'Exit and stop the application',
+            options: [_helpOption],
+            run: (c) async {
+              await CappConsole.progress(
+                "Bye bye!",
+                () async {
+                  await stop(force: true);
+                  await Future.delayed(Duration(seconds: 1), () {
+                    exit(0);
+                  });
+                },
+                type: CappProgressType.circle,
+              );
+            },
+          ),
           ...commands,
         ],
       );
@@ -689,7 +795,7 @@ class FinchApp {
 
     return _getCommandManager(args).processWhile(
       initArgs: args,
-      promptLabel: 'Finch> ',
+      appLabel: () => 'Finch> ',
     );
   }
 
@@ -837,6 +943,16 @@ class FinchApp {
             await debugger?.sendToAll(
               {'message': 'Language updated'},
               path: 'update_languages',
+            );
+          }),
+          'update_template': SocketEvent(onMessage: (socket, data) async {
+            var res = await WidgetToDart(
+              config.widgetsPath,
+              fileExtention: config.widgetsType,
+            ).generate();
+            await debugger?.sendToAll(
+              {'message': 'Template updated: $res'},
+              path: 'update_template',
             );
           }),
           'restart': SocketEvent(onMessage: (socket, data) async {
@@ -1073,5 +1189,5 @@ class _Info {
   /// - MINOR: New features (backward compatible)
   /// - PATCH: Bug fixes (backward compatible)
   /// - PRERELEASE: Pre-release identifiers (alpha, beta, rc)
-  final String version = '1.0.4';
+  final String version = '1.1.3';
 }